feat: Notification 엔티티를 알림 전송 주체에 따라 두 개의 엔티티로 분리
- ServerNotification: 서버에서 사용자에게 보내는 알림
- UserNotification: 사용자간의 상호작용에 대한 알림
