--- src/main/java/today/seasoning/seasoning/notification/domain/NotificationRepository.java
@@ -1,6 +1,0 @@
-package today.seasoning.seasoning.notification.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface NotificationRepository extends JpaRepository<Notification, Long> {
-}
